+++ app/src/test/kotlin/com/maltaisn/notes/ui/edit/EditViewModelTest.kt
@@ -314,8 +314,8 @@ class EditViewModelTest {
             EditDateItem(dateFor("2020-03-30").time),
             EditTitleItem("title".e, true),
             EditContentItem("- item 1\n- item 2".e, true),
-            EditChipsItem(listOf(labelsRepo.requireLabelById(1),
-                labelsRepo.requireLabelById(2)))
+            EditChipsItem(listOf(notesRepo.requireNoteById(3).reminder!!,
+                labelsRepo.requireLabelById(1), labelsRepo.requireLabelById(2)))
         ), viewModel.editItems.getOrAwaitValue())
 
         assertLiveDataEventSent(viewModel.focusEvent, EditViewModel.FocusChange(2, 17, false))
@@ -455,8 +455,8 @@ class EditViewModelTest {
             EditItemItem("item 1".e, checked = false, editable = true, 0),
             EditItemItem("item 2".e, checked = false, editable = true, 1),
             EditItemAddItem,
-            EditChipsItem(listOf(labelsRepo.requireLabelById(1),
-                labelsRepo.requireLabelById(2))),
+            EditChipsItem(listOf(notesRepo.requireNoteById(3).reminder!!,
+                labelsRepo.requireLabelById(1), labelsRepo.requireLabelById(2))),
         ), viewModel.editItems.getOrAwaitValue())
         assertLiveDataEventSent(viewModel.focusEvent, EditViewModel.FocusChange(1, 12, true))
 
@@ -834,7 +834,8 @@ class EditViewModelTest {
             EditItemItem("item 1".e, checked = false, editable = true, 0),
             EditItemItem("item 2".e, checked = false, editable = true, 1),
             EditItemAddItem,
-            EditChipsItem(listOf(labelsRepo.requireLabelById(1), labelsRepo.requireLabelById(2))),
+            EditChipsItem(listOf(notesRepo.requireNoteById(3).reminder!!,
+                labelsRepo.requireLabelById(1), labelsRepo.requireLabelById(2))),
         ), viewModel.editItems.getOrAwaitValue())
     }
 
@@ -913,7 +914,8 @@ class EditViewModelTest {
             EditItemAddItem,
             EditCheckedHeaderItem(1),
             EditItemItem("item 1".e, checked = true, editable = true, 0),
-            EditChipsItem(listOf(labelsRepo.requireLabelById(1), labelsRepo.requireLabelById(2))),
+            EditChipsItem(listOf(notesRepo.requireNoteById(3).reminder!!,
+                labelsRepo.requireLabelById(1), labelsRepo.requireLabelById(2))),
         ), viewModel.editItems.getOrAwaitValue())
     }